refactor code to TelemetryService
