+++ src/extension.ts
@@ -1,37 +1,31 @@
 import * as vscode from "vscode";
-import { SegmentInitializer } from "./segmentInitializer";
 import { Logger } from "./utils/logger";
-import Analytics from "analytics-node";
 import { TelemetryEventQueue } from "./utils/telemetryEventQueue";
-import { TelemetryEvent } from "./interfaces/telemetryEvent";
+import { TelemetryService } from "./services/telemetryService";
 import { Reporter } from "./reporter";
-
-/* 
-  OPT_IN_STATUS == "true" if user Agreed
-  false when denied for telemetry data collection     
-*/
-const OPT_IN_STATUS = "redhat.telemetry.optInRequested";
 
 // this method is called when your extension is activated
 export function activate(context: vscode.ExtensionContext) {
   Logger.log('"vscode-commons" is now active!');
 
-  //  instance of initialized segment object
-  const analytics: Analytics | undefined = SegmentInitializer.initialize(
-    context
-  );
-
-  if (!getTelemetryEnabledConfig()) {
+  /* 
+    check for current status of telemetry true | false
+    initialize events queue to preserve events 
+  */
+  if (!TelemetryService.getTelemetryEnabledConfig()) {
     Logger.log("redhat.telemetry.enabled is false");
     TelemetryEventQueue.initialize();
   } else {
     Logger.log("redhat.telemetry.enabled is true");
   }
 
-  openTelemetryOptInDialogIfNeeded(context);
+  TelemetryService.openTelemetryOptInDialogIfNeeded(context);
 
   context.subscriptions.push(
-    vscode.commands.registerCommand("vscodeCommons.openWebPage", openWebPage)
+    vscode.commands.registerCommand(
+      "vscodeCommons.openWebPage",
+      TelemetryService.openWebPage
+    )
   );
 
   /* 
@@ -40,15 +34,28 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(onDidChangeTelemetryEnabled());
 
   /* 
-    test command to activate extension through command pallet
-    can be removed later 
+    test command to activate and  view telemetry status of 
+    extension through command pallet can be removed later 
   */
   context.subscriptions.push(
     vscode.commands.registerCommand("vscodeCommons.showTelemetryStatus", () => {
       vscode.window.showInformationMessage(
-        `Telemetry Enabled: ${getTelemetryEnabledConfig()}`
+        `Red Hat Telemetry Enabled: ${TelemetryService.getTelemetryEnabledConfig()}`
       );
     })
+  );
+
+  // This command exists only for testing purposes. Should delete later.
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
+      "vscodeCommons.clearStateAndSettings",
+      () => {
+        context.globalState.update("OPT_IN_STATUS", undefined);
+        vscode.workspace
+          .getConfiguration("redhat.telemetry")
+          .update("enabled", undefined, true);
+      }
+    )
   );
 
   /* 
@@ -56,14 +63,14 @@ export function activate(context: vscode.ExtensionContext) {
     These APIs can be used in by other extensions as exports
     Please view INTEGRATION.md for more details
   */
-  if (analytics) {
-    Reporter.setAnalytics(analytics);
-    // export to other extensions
-    return Promise.resolve({
-      telemetryData,
-      viewMessage,
-    });
-  }
+  // if (analytics) {
+  //   Reporter.setAnalytics(analytics);
+  //   // export to other extensions
+  //   return Promise.resolve({
+  //     telemetryData,
+  //     viewMessage,
+  //   });
+  // }
 }
 
 /* Basic test api MUST BE REMOVED LATER */
@@ -71,77 +78,6 @@ function viewMessage(msg: string) {
   vscode.window.showInformationMessage(
     `Msg Received in Vscode-common:  ${msg}`
   );
-}
-
-/* 
-  Collects telemetry data and pushes to a queue when not opted in
-  and to segment when user has enabled telemetry 
-*/
-function telemetryData(e: TelemetryEvent) {
-  Logger.log("Event received:");
-  Logger.log(e.extensionName);
-  if (getTelemetryEnabledConfig()) {
-    // dequeue the existing queue
-    TelemetryEventQueue.getQueue()?.length &&
-      Reporter.reportQueue(TelemetryEventQueue.getQueue());
-    // report event to segment
-    Reporter.report(e);
-  } else {
-    TelemetryEventQueue.addEvent(e);
-  }
-}
-
-/* 
-	Checks the current telemetry configuration(package.json) 
-	returns "true if telemetry is enabled 
-			"false" if it is not-enable or undefined(default value)"
-*/
-function getTelemetryEnabledConfig(): boolean {
-  return vscode.workspace
-    .getConfiguration("redhat.telemetry")
-    .get<boolean>("enabled", false);
-}
-
-function openTelemetryOptInDialogIfNeeded(context: vscode.ExtensionContext) {
-  const optInRequested: boolean | undefined = context.globalState.get(
-    OPT_IN_STATUS
-  );
-  Logger.log(`optInRequested is: ${optInRequested}`);
-
-  if (!optInRequested) {
-    const privacyUrl: string =
-      "https://github.com/redhat-developer/vscode-commons/wiki/Usage-reporting";
-    const command: string = "vscodeCommons.openWebPage";
-    const message: string = `Red Hat would like to collect some usage data from its extensions. 
-                            [Read our privacy statement](command:${command}?"${privacyUrl}").`;
-
-    vscode.window
-      .showInformationMessage(message, "Accept", "Deny")
-      .then((selection) => {
-        if (!selection) {
-          //close was chosen. Ask next time.
-          return;
-        }
-
-        context.globalState.update(OPT_IN_STATUS, true);
-
-        let optIn: boolean = selection === "Accept";
-        updateTelemetryEnabledConfig(optIn);
-      });
-  }
-}
-
-function updateTelemetryEnabledConfig(value: boolean): Thenable<void> {
-  return vscode.workspace
-    .getConfiguration("redhat.telemetry")
-    .update("enabled", value, true);
-}
-
-/*
-  open vscode-commons privacy page. Used in opt in notification(openTelemetryOptInDialogIfNeeded)
-*/
-function openWebPage(url: string) {
-  vscode.commands.executeCommand("vscode.open", vscode.Uri.parse(url));
 }
 
 /* 
@@ -161,7 +97,7 @@ function onDidChangeTelemetryEnabled(): vscode.Disposable {
       if (!e.affectsConfiguration("redhat.telemetry.enabled")) {
         return;
       }
-      if (getTelemetryEnabledConfig()) {
+      if (TelemetryService.getTelemetryEnabledConfig()) {
         Logger.log("redhat.telemetry.enabled set to true");
       } else {
         Logger.log("redhat.telemetry.enabled set to false");
