+++ src/utils/segmentInitializer.ts
